+++ src/client/common/components/index.js
@@ -1,0 +1,15 @@
+const AsyncButton = require('./async-button');
+const DropDown = require('./drop-down');
+const FlatButton = require('./flat-button');
+const Popover = require('./popover');
+const Popup = require('./popup');
+const TextTooltip = require('./text-tooltip');
+
+module.exports = {
+  AsyncButton,
+  DropDown,
+  FlatButton,
+  Popover,
+  Popup,
+  TextTooltip
+};
